requirements.txt
requests>=2.28.2
bs4>=0.0.1
beautifulsoup4>=4.11.1
cinemagoer>=2022.12.27
python-telegram-bot>=13.10
